requirements.txt
fastapi
uvicorn
scikit-learn
python-multipart
pypdf
python-docx
requests